Revert "WCLI screenshot"
