+++ .gitignore
@@ -4,6 +4,4 @@
 .vscode/launch.json
 .vscode/ipch
 .vscode/settings.json
-assets
-tools/*.raw
-tools/*.png
+
