+++ lib/cli/cli.cpp
@@ -8,7 +8,8 @@
 
 #ifndef DISABLE_CLI
 #include "cli.hpp"
-#include "utils.h"
+#include "storage.hpp"
+#include "tft.hpp"
 
 const char logo[] =
 "\r\n"
@@ -23,7 +24,39 @@ const char logo[] =
 ""
 ;
 
+// Capture the screenshot and save it to the SD card
+void captureScreenshot(const char* filename, Stream *response) {
+  File file = SD.open(filename, FILE_WRITE);
+  if (!file) {
+    response->println("Failed to open file for writing");
+    return;
+  }
 
+  // Allocate memory to store the screen data
+  response->printf("Allocating memory for: %ix%i image\r\n",tft.width(),tft.height());
+  uint8_t* buffer = (uint8_t*)ps_malloc(tft.width() * tft.height() * sizeof(uint8_t));
+  if (!buffer) {
+    response->println("Failed to allocate memory for buffer");
+    file.close();
+    return;
+  }
+
+  // Read the screen data into the buffer
+  tft.readRect(0, 0, tft.width(), tft.height(), buffer);
+
+  // Write the buffer data to the file
+  for (int y = 0; y < tft.height(); y++) {
+    for (int x = 0; x < tft.width(); x++) {
+      uint16_t color = buffer[y * tft.width() + x];
+      file.write((const uint8_t*)&color, sizeof(color));
+    }
+  }
+
+  // Clean up
+  free(buffer);
+  file.close();
+  response->println("Screenshot saved");
+}
 
 void wcli_reboot(char *args, Stream *response) {
   ESP.restart();
@@ -39,38 +72,7 @@ void wcli_clear(char *args, Stream *response){
 }
 
 void wcli_scshot(char *args, Stream *response){
-  Pair<String, String> operands = wcli.parseCommand(args);
-  String ip = operands.first();
-  uint16_t port = operands.second().toInt();
- 
-  if (ip.isEmpty()){
-    response->println("Saving to SD..");
-    captureScreenshot(SCREENSHOT_TEMP_FILE, response);
-    response->println("Note: is possible to send it to a PC using: scshot ip port");
-  }
-  else {
-    response->printf("Sending screenshot to %s:%i..\r\n", ip.c_str(), port);
-    captureScreenshot(SCREENSHOT_TEMP_FILE, response);
-    captureScreenshot(SCREENSHOT_TEMP_FILE, ip.c_str(), port);
-  }
-}
-
-void initRemoteShell(){
-#ifndef DISABLE_CLI_TELNET 
-  if (wcli.isTelnetEnable()) wcli.shellTelnet->attachLogo(logo);
-#endif
-}
-
-void initShell(){
-  wcli.shell->attachLogo(logo);
-  wcli.setSilentMode(true);
-  wcli.disableConnectInBoot();
-  // Main Commands:
-  wcli.add("reboot", &wcli_reboot, "\tperform a ESP32 reboot");
-  wcli.add("info", &wcli_info, "\t\tget device information");
-  wcli.add("clear", &wcli_clear, "\t\tclear shell");
-  wcli.add("scshot", &wcli_scshot, "\tscreenshot to SD or sending a PC");
-  wcli.begin("IceNav");
+  captureScreenshot("/screenshot.raw",response);
 }
 
 /**
@@ -82,9 +84,22 @@ void initCLI() {
     while (!Serial){};
     delay(1000);
   #endif
-  log_v("init CLI");
-  initShell(); 
-  initRemoteShell();
+  Serial.println("init CLI");
+  wcli.shell->attachLogo(logo);
+  wcli.setSilentMode(true);
+  wcli.disableConnectInBoot();
+  // Main Commands:
+  wcli.add("reboot", &wcli_reboot, "\tperform a ESP32 reboot");
+  wcli.add("info", &wcli_info, "\t\tget device information");
+  wcli.add("clear", &wcli_clear, "\t\tclear shell");
+  wcli.add("scshot", &wcli_scshot, "\ttake screen shot");
+  
+  wcli.begin("IceNav");
+
+#ifndef DISABLE_CLI_TELNET 
+  wcli.enableTelnet();
+  wcli.shellTelnet->attachLogo(logo);
+#endif
 }
 
 #endif
