+++ platformio.ini
@@ -24,7 +24,7 @@ monitor_filters =
 	esp32_exception_decoder
 extra_scripts = pre:prebuild.py
 build_flags = 
-	-D CORE_DEBUG_LEVEL=0
+	-D CORE_DEBUG_LEVEL=5
 	-D USE_LINE_BUFFER=1
 	-D DISABLE_RADIO=1
 	-D BAUDRATE=115200
